+++ changelog.md
@@ -1,6 +1,5 @@
 [![Downloads](https://img.shields.io/github/downloads/probablyraging/steam-game-idler/1.5.27/total?style=for-the-badge&logo=github&color=137eb5)](https://github.com/probablyraging/steam-game-idler/releases/download/1.5.27/Steam.Game.Idler_1.5.27_x64_en-US.msi)
 
 ## Changelog
-- Fixed some UI blocking issues when running specific Tauri commands
-- Improved handling for unlocking/locking achievements, especially when done in bulk
-- Refactored the `SteamUtility` lib
+- Replaced the individual `Update` buttons with a single `Save changes` button on the `Achievement > Statistics` tab
+  - (#36) by @Jesewe